requirements.txt
openai
discord.py[voice]
tiktoken
PyNaCl